fix(learn): keep the API client out of the catalogue's import path
The scope-tours checker runs the catalogue under tsx, outside vite, and
access.ts pulled in the API client for its hook, which reads
import.meta.env at module load: `Cannot read properties of undefined
(reading 'VITEST')`. The hook moves to its own file; access.ts stays
the pure scope rules the catalogue needs.

CS-267
